net模块-定义包结构、proto包处理handler

+++ net/src/main/java/com/awake/net/packet/IPacket.java
@@ -9,6 +9,6 @@ package com.awake.net.packet;
  **/
 public interface IPacket {
 
-    short protocolId();
+    int protocolId();
 
 }
